Server/Scripts: Add some missing overrides.

+++ src/server/game/Entities/Pet/Pet.cpp
@@ -217,7 +217,7 @@ bool Pet::LoadPetFromDB(Player* owner, uint32 petEntry, uint32 petnumber, bool c
             break;
         default:
             if (!IsPetGhoul())
-                SF_LOG_ERROR("entities.pet", "Pet have incorrect type (%u) for pet loading.", getPetType());
+                SF_LOG_ERROR("entities.pet", "Pet have incorrect type (%u) for pet loading.", uint8(getPetType()));
             break;
     }
